Task 1 consisting of 3 files: "main", "user_input" and "clip_raster_file"
Main.py is the main program that will call the various defined functions to complete the tasks.
User_input.py contains a defined function to determine whether the user input location lies within the MBR as well as the shapefile of the isle of wight.
Clip_raster_file.py contains a defined function that clips out the raster data within 5km from the user's location, transforms the clipped data into the correct coordinate reference system, and saves it as a new file "output_SZ.asc".